--- profit/uq/__init__.py
@@ -1,2 +1,0 @@
-from .uq import *
-from .backend import *
